Fix of essential food calculations in registered alimentation for deployed server

+++ app/src/main/java/study/snacktrackmobile/viewmodel/RegisteredAlimentationViewModel.kt
@@ -26,6 +26,7 @@ import study.snacktrackmobile.data.services.AiApiService
 import study.snacktrackmobile.data.services.AiShoppingRequest
 import study.snacktrackmobile.data.storage.TokenStorage
 import study.snacktrackmobile.presentation.ui.state.SummaryBarState
+import java.time.LocalDate
 
 class RegisteredAlimentationViewModel(
     private val repository: RegisteredAlimentationRepository,
@@ -51,7 +52,6 @@ class RegisteredAlimentationViewModel(
     private var loadMealsJob: Job? = null
 
     // 🔹 POMOCNICZA FUNKCJA: Wyciąganie wagi ze stringa (np. "330ml" -> 330.0)
-    // To ta sama logika co w FoodViewModel i ProductDetailsScreen
     private fun extractWeight(quantityStr: String?): Float? {
         if (quantityStr == null) return null
         val regex = Regex("(\\d+(?:\\.\\d+)?)\\s*(g|ml|l|kg)", RegexOption.IGNORE_CASE)
@@ -164,7 +164,7 @@ class RegisteredAlimentationViewModel(
         }
     }
 
-    // 🔹 FIX: Zaktualizowana logika mapowania z parsowaniem wagi
+    // 🔹 FIX: Logika mapowania z użyciem Double dla makroskładników
     private fun mapToUi(input: List<RegisteredAlimentationResponse>): List<Meal> {
         return input
             .groupBy { normalizeMealName(it.mealName ?: "Other") }
@@ -179,9 +179,10 @@ class RegisteredAlimentationViewModel(
                             val ef = ing.essentialFood
                             val api = ing.essentialApi
 
-                            val baseCal = (ef?.calories ?: api?.calorie?.toFloat() ?: 0f).toDouble()
+                            // Użycie Double dla makroskładników
+                            val baseCal = (ef?.calories ?: api?.calorie?.toDouble() ?: 0.0)
 
-                            // 🔹 FIX: Parsowanie wagi składnika przepisu
+                            // Parsowanie wagi składnika przepisu
                             val determinedWeight = ef?.defaultWeight
                                 ?: api?.defaultWeight
                                 ?: extractWeight(api?.quantity)
@@ -208,10 +209,10 @@ class RegisteredAlimentationViewModel(
                         val ef = entry.essentialFood
                         val api = entry.mealApi
 
-                        val baseCal = (ef?.calories ?: api?.calorie?.toFloat() ?: 0f).toDouble()
+                        // Użycie Double dla makroskładników
+                        val baseCal = (ef?.calories ?: api?.calorie?.toDouble() ?: 0.0)
 
-                        // 🔹 FIX: Parsowanie wagi dla pojedynczego produktu
-                        // Teraz szukamy w quantity ("330ml") i servingSizeUnit, zanim użyjemy fallbacku 100f
+                        // Parsowanie wagi dla pojedynczego produktu
                         val determinedWeight = ef?.defaultWeight
                             ?: api?.defaultWeight
                             ?: extractWeight(api?.quantity)
